requirements.txt
pandas==2.2.2
pyyaml==6.0.2
python-dotenv==1.0.1
pytrends==4.9.2
praw==7.7.1
snscrape==0.7.0.20230622
requests==2.32.3
beautifulsoup4==4.12.3
